requirements.txt
accelerate>=0.23.0
datasets>=2.14.6
deepspeed>=0.12.2
einops>=0.6.1
evaluate==0.4.0
huggingface-hub>=0.14.1<1.0
ninja>=1.11.1
packaging>=23.0
parameterized>=0.9.0
peft>=0.6.1
protobuf<=3.20.2
safetensors>=0.3.3
tensorboard
transformers>=4.35.0
trl>=0.7.4
jinja2>=3.0.0
tqdm>=4.64.1
flash-attn>=2.1.0
pynvml>=11.4.0

# optional
galore-torch

# unsloth
# with NV pytorch container install -> pip install git+https://github.com/unslothai/unsloth.git --no-deps